implemented pthread_attr to get stack top and bottom

+++ inc/linkedList.h
@@ -1,6 +1,8 @@
 #ifndef _LINKED_LIST_H_
 #define _LINKED_LIST_H_
 
+#include <stddef.h>
+
 typedef struct llNode {
     long threadID;
     long stack_top;
@@ -9,9 +11,9 @@ typedef struct llNode {
 } llNode;
 
 
-void ll_insertNode(llNode** head, long threadID, long stack_top, long stack_bottom);
+void ll_insertNode(llNode** head, size_t threadID, size_t stack_top, size_t stack_bottom);
 void ll_destroy(llNode** head);
-void ll_removeNode(llNode** head, long threadID);
+void ll_removeNode(llNode** head, size_t threadID);
 
 
 #endif
